Remove `pest_meta` dependency on `quote`
As `quote::Ident` is the only thing from `quote` used by `pest_meta`,
the dependency on `quote` is overkill. Replace uses of `quote::Ident` by
`String` and remove the dependency on `quote`.

As `pest_derive` relies on `quote::Ident` for the code generation part,
it is OK to keep `quote` as a dependency. Create the required
`quote::Ident` instances from the identifier names which are now
provided by `pest_meta` as `String`.