Refactored Error.
The former Error type kept reference to the input. This meant that
it was impossible for an Error to be fully owned, resulting in a
hard-to-use type that was incompatible with failure. This commit
Refactors the error code, implements a friendlier API, promotes
error to a public mod, and transforms Error into an owned type.
Fixes #204.