Allow infix operators to prefix expressions
Currently, writing an expression with choice or sequence operators that is
spread across multiple lines requires either appending the operator to each line
except the last, or prefixing each line with the operator except the first. Both
are awkward to read and use:

```pest
number = @{ '0'..'9'+ }
prefix_expr = {
    number ~ "+" ~ number
    | number ~ "-" ~ number
    | number
}
suffix_expr = {
    number ~ "+" ~ number |
    number ~ "-" ~ number |
    number
}
```

This commit seeks to solve the latter issue by allowing a single infix operator
to be placed at the beginning of an expression without affecting the meaning of
the expression:

```pest
number = @{ '0'..'9'+ }
expr = {
    | number ~ "+" ~ number
    | number ~ "-" ~ number
    | number
}
```

I had to remove the version requirement from the derive Cargo.toml file, as it
wouldn't use the newly bootstrapped parser otherwise. (Frankly, I'm not sure
exactly how to go about the bootstrap system.)